Split workspace proxy middleware into separate file
