+++ apps/prairielearn/src/server.ts
@@ -26,8 +26,6 @@ import express, {
   type Response,
 } from 'express';
 import asyncHandler from 'express-async-handler';
-import { createProxyMiddleware } from 'http-proxy-middleware';
-import type * as httpProxyMiddleware from 'http-proxy-middleware';
 import _ from 'lodash';
 import multer from 'multer';
 import onFinished from 'on-finished';
@@ -37,7 +35,6 @@ import { v4 as uuidv4 } from 'uuid';
 import yargsParser from 'yargs-parser';
 
 import { cache } from '@prairielearn/cache';
-import * as error from '@prairielearn/error';
 import { flashMiddleware } from '@prairielearn/flash';
 import { addFileLogging, logger } from '@prairielearn/logger';
 import * as migrations from '@prairielearn/migrations';
@@ -68,7 +65,6 @@ import { featuresMiddleware } from './lib/features/middleware.js';
 import { isEnterprise } from './lib/license.js';
 import * as lifecycleHooks from './lib/lifecycle-hooks.js';
 import * as load from './lib/load.js';
-import { LocalCache } from './lib/local-cache.js';
 import { APP_ROOT_PATH, REPOSITORY_ROOT_PATH } from './lib/paths.js';
 import * as serverJobs from './lib/server-jobs.js';
 import { PostgresSessionStore } from './lib/session-store.js';
@@ -79,6 +75,7 @@ import * as workspace from './lib/workspace.js';
 import { markAllWorkspaceHostsUnhealthy } from './lib/workspaceHost.js';
 import { enterpriseOnly } from './middlewares/enterpriseOnly.js';
 import staticNodeModules from './middlewares/staticNodeModules.js';
+import { makeWorkspaceProxyMiddleware } from './middlewares/workspaceProxy.js';
 import * as news_items from './news_items/index.js';
 import * as freeformServer from './question-servers/freeform.js';
 import * as sprocs from './sprocs/index.js';
@@ -277,35 +274,6 @@ export async function initExpress(): Promise<Express> {
     upload.single('file'),
   );
 
-  /**
-   * Function to strip "sensitive" cookies from requests that will be proxied
-   * to workspace hosts.
-   */
-  function stripSensitiveCookies(proxyReq: http.ClientRequest) {
-    const cookies = proxyReq.getHeader('cookie');
-    if (!cookies) return;
-
-    const items = (cookies as string).split(';');
-    const filteredItems = items.filter((item) => {
-      const name = item.split('=')[0].trim();
-      return (
-        name !== 'pl_authn' &&
-        name !== 'pl2_authn' &&
-        name !== 'pl_assessmentpw' &&
-        name !== 'pl2_assessmentpw' &&
-        name !== 'connect.sid' &&
-        name !== 'prairielearn_session' &&
-        name !== 'pl2_session' &&
-        // The workspace authz cookies use a prefix plus the workspace ID, so
-        // we need to check for that prefix instead of an exact name match.
-        !name.startsWith('pl_authz_workspace_') &&
-        !name.startsWith('pl2_authz_workspace_')
-      );
-    });
-
-    proxyReq.setHeader('cookie', filteredItems.join(';'));
-  }
-
   // Collect metrics on workspace proxy sockets. Note that this only tracks
   // outgoing sockets (those going to workspaces). Incoming sockets are tracked
   // globally for the entire server.
@@ -313,81 +281,6 @@ export async function initExpress(): Promise<Express> {
   const workspaceProxySocketActivityMetrics = new SocketActivityMetrics(meter, 'workspace-proxy');
   workspaceProxySocketActivityMetrics.start();
 
-  // proxy workspaces to remote machines
-  const workspaceUrlRewriteCache = new LocalCache(config.workspaceUrlRewriteCacheMaxAgeSec);
-  const workspaceProxyOptions: httpProxyMiddleware.Options = {
-    target: 'invalid',
-    ws: true,
-    pathRewrite: async (path) => {
-      try {
-        const match = path.match('/pl/workspace/([0-9]+)/container/(.*)');
-        if (!match) throw new Error(`Could not match path: ${path}`);
-        const workspace_id = parseInt(match[1]);
-        let workspace_url_rewrite = workspaceUrlRewriteCache.get(workspace_id);
-        if (workspace_url_rewrite == null) {
-          const sql =
-            'SELECT q.workspace_url_rewrite' +
-            ' FROM questions AS q' +
-            ' JOIN variants AS v ON (v.question_id = q.id)' +
-            ' WHERE v.workspace_id = $workspace_id;';
-          const result = await sqldb.queryOneRowAsync(sql, { workspace_id });
-          workspace_url_rewrite = result.rows[0].workspace_url_rewrite;
-          if (workspace_url_rewrite == null) workspace_url_rewrite = true;
-          workspaceUrlRewriteCache.set(workspace_id, workspace_url_rewrite);
-        }
-        if (!workspace_url_rewrite) {
-          return path;
-        }
-        const pathSuffix = match[2];
-        const newPath = '/' + pathSuffix;
-        return newPath;
-      } catch (err) {
-        logger.error(`Error in pathRewrite for path=${path}: ${err}`);
-        return path;
-      }
-    },
-    logLevel: 'silent',
-    logProvider: (_provider) => logger,
-    router: async (req) => {
-      const match = req.url.match(/^\/pl\/workspace\/([0-9]+)\/container\//);
-      if (!match) throw new Error(`Could not match URL: ${req.url}`);
-
-      const workspace_id = match[1];
-      const result = await sqldb.queryZeroOrOneRowAsync(
-        "SELECT hostname FROM workspaces WHERE id = $workspace_id AND state = 'running';",
-        { workspace_id },
-      );
-
-      if (result.rows.length === 0) {
-        // If updating this message, also update the message our Sentry
-        // `beforeSend` handler.
-        throw new error.HttpStatusError(404, 'Workspace is not running');
-      }
-
-      return `http://${result.rows[0].hostname}/`;
-    },
-    onProxyReq: (proxyReq) => {
-      stripSensitiveCookies(proxyReq);
-    },
-    onProxyReqWs: (proxyReq) => {
-      stripSensitiveCookies(proxyReq);
-    },
-    onError: (err, req, res) => {
-      logger.error(`Error proxying workspace request: ${err}`, {
-        err,
-        url: req.url,
-        originalUrl: req.originalUrl,
-      });
-      // Check to make sure we weren't already in the middle of sending a
-      // response before replying with an error 500
-      if (res && !res.headersSent) {
-        res.status?.((err as any).status ?? 500)?.send?.('Error proxying workspace request');
-      }
-    },
-  };
-  const workspaceProxy = createProxyMiddleware((pathname) => {
-    return !!pathname.match('/pl/workspace/([0-9])+/container/');
-  }, workspaceProxyOptions);
   const workspaceAuthRouter = express.Router();
   workspaceAuthRouter.use([
     // We use a short-lived cookie to cache a successful
@@ -417,7 +310,7 @@ export async function initExpress(): Promise<Express> {
       workspaceProxySocketActivityMetrics.addSocket(req.socket);
       next();
     },
-    workspaceProxy,
+    makeWorkspaceProxyMiddleware(),
   ]);
 
   app.use((req, res, next) => {
